handle case where mock server returns warning or error logs. The example which caused this to be invesitgated and then fixed is when you use an incorrect format for a content type response. Consider the following text/plain response
text/plain:
  example: |
  title: "Mocked Text title"
  completed: true
  order: 13
  url: "http://mockedURL.com"

When calling the endpoint, mockserver would emit this warning:
{"level":"warning","msg":"only string schema is supported for
'text/plain' content
type","requestId":"305aeb23-0b99-4352-b662-167f2c22562f","time":"2022-07-19T11:32:44Z"}.
Firstly, this should be displayed to the user. Secondly, the
decorateLogEntry function expects the log entry to be well formed so it
can format it correctly.

When a warning or error such as this is emitted, we log it out as a
warning and not format it